backend/pyproject.toml
[project]
name = "tg-paid-subscriber-service"
version = "0.1.0"
description = "Production-grade Telegram paid subscriber service with FastAPI"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "black>=25.9.0",
    "fastapi>=0.121.0",
    "granian>=2.5.7",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.4",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "ruff>=0.14.3",
    "stripe>=13.2.0",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov=routers",
    "--cov=config",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests that don't require external dependencies",
    "integration: Integration tests that may require MongoDB or external services",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["app", "routers", "config"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused

[tool.black]
line-length = 100
target-version = ['py313']
include = '\.pyi?$'

